+++ trace-replay/Makefile
@@ -1,2 +1,4 @@
-main:
-	g++ -fopenmp -O3 -o main trace-replay-kernel.cpp
+trace-replay-kernel.o: trace-replay-kernel.cpp trace-replay-kernel.h
+	g++ -fopenmp -c -O3 -o trace-replay-kernel.o
+main: trace-replay-kernel.o main.cpp
+	g++ -fopenmp -O3 -o main trace-replay-kernel.o main.cpp
