+++ trace-replay/trace-replay-kernel.cpp
@@ -4,6 +4,8 @@
 #include <cstdlib>
 #include <cstdio>
 
+#include "trace-replay-kernel.h"
+
 // The first byte of trace_entry is an enum
 //   0 -> read,  size 4
 //   1 -> read,  size 8
@@ -34,11 +36,11 @@
 typedef uint64_t trace_entry;
 
 int rw_sz(trace_entry t) {
-    return t >> 56;
+    return t >> ADDR_BITS;
 }
 
 int addr(trace_entry t) {
-    return t & 0xFF'FF'FF'FF'FF'FF'FFLL;
+    return t & ADDR_MASK;
 }
 
 void trace_replay_kernel(trace_entry *tr, long len, void *mem, double *local) {
@@ -67,36 +69,4 @@ void trace_replay_kernel(trace_entry *tr, long len, void *mem, double *local) {
     }
 }
 
-}
-
-// Helpers for creating the trace
-long to_rwsz(long long a) {
-    return a << 56;
-}
-
-long to_addr(long long a) {
-    return a & 0xFF'FF'FF'FF'FF'FF'FFLL;
-}
-
-int main() {
-    assert(sizeof(long) == 8);
-
-    // Create a trace with 8 entries
-    trace_entry *tr = (trace_entry*)malloc(sizeof(trace_entry) * 8);
-
-    tr[0] = to_rwsz(0) | to_addr(0x0);  // Read  4
-    tr[1] = to_rwsz(1) | to_addr(0x8);  // Read  8
-    tr[2] = to_rwsz(2) | to_addr(0xC);  // Write 4
-    tr[3] = to_rwsz(3) | to_addr(0x18); // Write 8
-
-    tr[4] = to_rwsz(3) | to_addr(0x30); // Write 8
-    tr[5] = to_rwsz(2) | to_addr(0x34); // Write 4
-    tr[6] = to_rwsz(1) | to_addr(0x40); // Read  8
-    tr[7] = to_rwsz(0) | to_addr(0x44); // Read  4
-
-    double *local = (double*)malloc(sizeof(double) * GAP_SIZE * omp_get_max_threads());
-    double *mem   = (double*)malloc(sizeof(char)   * 0x64);
-
-    trace_replay_kernel(tr, 8, mem, local);
-    printf("All done!\n");
 }
